requirements.txt
python-dotenv
streamlit
openai
langchain
PyPDF2
faiss-cpu
altair
tiktoken
Flask

# Tests
mock